Refactor payee splitter into package
