+++ payee_splitter/__init__.py
@@ -1,0 +1,3 @@
+from .core import split_payee_desc_block
+
+__all__ = ["split_payee_desc_block"]
