+++ payee_splitter/constants.py
@@ -1,0 +1,85 @@
+STOPWORDS = {
+    "MERCHANT",
+    "OFFICE",
+    "SUPPLIES",
+    "EXPENSE",
+    "FEE",
+    "FEES",
+    "PAYMENT",
+    "RE",
+    "RE:",
+    "TOTAL",
+    "REIMBURSEMENT",
+    "REIMBURSE",
+    "PERFORMANCE",
+    "CONTRACT",
+    "RENTAL",
+    "PROGRAM",
+    "TRAINING",
+    "PER",
+    "DIEM",
+    "INVOICE",
+    "PROFESSIONAL",
+    "TUITION",
+}
+
+KNOWN_PREFIXES = [
+    "ALAMEDA COUNTY FIRE DEPARTMENT",
+    "BAY AREA NEWS GROUP",
+    "DIEGO TRUCK REPAIR",
+    "L.N. CURTIS & SONS",
+    "J & O'S COMMERCIAL TIRE CENTER",
+    "MUNICIPAL POOLING AUTHORITY",
+    "KAISER FOUNDATION HEALTH PLAN",
+    "EAST BAY REGIONAL COMMUNICATIONS SYSTEM",
+    "CONTRA COSTA HEALTH SERVICES",
+    "GHIRARDELLI ASSOCIATES",
+    "FLOCK SAFETY",
+    "PERS",
+]
+
+SUFFIXES = {
+    "LLP",
+    "LLC",
+    "INC",
+    "CORP",
+    "CORPORATION",
+    "CO",
+    "COMPANY",
+    "LTD",
+    "ASSOCIATES",
+    "SUPPLY",
+    "SERVICE",
+    "SERVICES",
+    "MANAGEMENT",
+    "ELECTRIC",
+}
+
+MONTHS = {
+    "JAN",
+    "JANUARY",
+    "FEB",
+    "FEBRUARY",
+    "MAR",
+    "MARCH",
+    "APR",
+    "APRIL",
+    "MAY",
+    "JUN",
+    "JUNE",
+    "JUL",
+    "JULY",
+    "AUG",
+    "AUGUST",
+    "SEP",
+    "SEPT",
+    "SEPTEMBER",
+    "OCT",
+    "OCTOBER",
+    "NOV",
+    "NOVEMBER",
+    "DEC",
+    "DECEMBER",
+}
+
+PREFIX_SET = {p.upper() for p in KNOWN_PREFIXES}
